refactoring

+++ app/src/main/java/ru/solandme/washwait/MainActivity.java
@@ -289,8 +289,8 @@ public class MainActivity extends AppCompatActivity implements SwipeRefreshLayou
     }
 
     private void startWashCarActivity() {
-      float lat = sharedPref.getFloat("lat", ForecastService.DEFAULT_LATITUDE);
-      float lon = sharedPref.getFloat("lon", ForecastService.DEFAULT_LONGITUDE);
+      float lat = sharedPref.getFloat("lat", (float) ForecastService.DEFAULT_LATITUDE);
+      float lon = sharedPref.getFloat("lon", (float) ForecastService.DEFAULT_LONGITUDE);
         Intent intent = new Intent(MainActivity.this, MapActivity.class);
         intent.putExtra("lat", lat);
         intent.putExtra("lon", lon);
